martello i toth - przebudowa brute'a, dodanie procedury L2, duzo smieci

+++ src/common/Constants.java
@@ -4,5 +4,5 @@ package common;
  * Created by michto on 24.10.15.
  */
 public class Constants {
-    public static final int MAX_CAPACITY_OF_BOX = 1;
+    public static final double MAX_CAPACITY_OF_BOX = 1;
 }
